+++ romancal/orientation/tests/test_set_telescope_pointing.py
@@ -15,7 +15,7 @@ import asdf
 import roman_datamodels as rdm
 from astropy.time import Time
 
-from romancal.lib import engdb_mast, engdb_tools
+from romancal.lib.engdb import engdb_mast, engdb_tools
 from romancal.orientation import set_telescope_pointing as stp
 
 # Ensure that `set_telescope_pointing` logs.
